+++ Weather/ViewModels/CurrentViewModel.swift
@@ -45,6 +45,13 @@ class CurrentViewModel: ObservableObject {
   }
 
   func saveToDatabase(_ current: Current) {
+    guard let context = repository?.context else {return}
+    if let currentEntity = current.toNSManagedObject(in: context) {
+      do {
+        try repository?.create(currentEntity)
+      } catch {
 
+      }
+    }
   }
 }
