+++ Weather/Weather.xcdatamodeld/Weather 3.xcdatamodel/contents
@@ -24,6 +24,6 @@
         <attribute name="icon" optional="YES" attributeType="String"/>
         <attribute name="main" optional="YES" attributeType="String"/>
         <attribute name="main_description" optional="YES" attributeType="String"/>
-        <relationship name="current" maxCount="1" deletionRule="Nullify" destinationEntity="Current" inverseName="weather" inverseEntity="Current"/>
+        <relationship name="current" optional="YES" maxCount="1" deletionRule="Nullify" destinationEntity="Current" inverseName="weather" inverseEntity="Current"/>
     </entity>
 </model>
