+++ WeatherTests/Models/MapAnnotationTests.swift
@@ -11,12 +11,14 @@ import CoreLocation
 
 final class MapAnnotationTests: XCTestCase {
 
+  var sut: MapAnnotation!
+
   func test_mapannotation_has_name_coordinates() {
 
-    let annotation = MapAnnotation(name: "name", coordinate: CLLocationCoordinate2D(latitude: 0.01, longitude: 31.2))
+    sut = MapAnnotation(name: "name", coordinate: CLLocationCoordinate2D(latitude: 0.01, longitude: 31.2))
 
-    XCTAssertNotNil(annotation.name)
-    XCTAssertNotNil(annotation.coordinate)
+    XCTAssertNotNil(sut.name)
+    XCTAssertNotNil(sut.coordinate)
   }
 
 }
